Currently building with Vector3f (float) - need to see if double can be used but this builds. I now need to determine which other vecmath types work here; specifically TexCoord2f and Quat4f

+++ src/main/java/com/cleversoap/jota/entity/MeshEntity.java
@@ -1,19 +1,19 @@
 package com.cleversoap.jota.entity;
 
 //import com.hackoeur.jglm.Vec2;
-//import com.hackoeur.jglm.Vector3d;
+//import com.hackoeur.jglm.Vector3f;
 import com.hackoeur.jglm.Vec4;
 
 import javax.vecmath.Vector2d;
-import javax.vecmath.Vector3d;
+import javax.vecmath.Vector3f;
 
 import java.util.List;
 import java.util.ArrayList;
 
 public class MeshEntity extends Entity
 {
-    protected List<Vector3d> _vertices;
-    protected List<Vector3d> _normals;
+    protected List<Vector3f> _vertices;
+    protected List<Vector3f> _normals;
 
     // TODO: Look at using the TexCoord2d type
     protected List<Vector2d> _texCoords;
@@ -22,13 +22,13 @@ public class MeshEntity extends Entity
     {
         super();
 
-        _vertices = new ArrayList<Vector3d>();
-        _normals = new ArrayList<Vector3d>();
+        _vertices = new ArrayList<Vector3f>();
+        _normals = new ArrayList<Vector3f>();
         _texCoords = new ArrayList<Vector2d>();
     }
 
     @Override
-    public void setPosition(Vector3d value)
+    public void setPosition(Vector3f value)
     {
         super.setPosition(value);
         // TODO: Update mesh matricies
